dom/html/test/imports/test_element_upgrade.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1081107
-->
<head>
  <title>Test for Bug 1081107</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <meta http-equiv="Content-Type" content="text/html;charset=utf-8">
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1081107">Mozilla Bug 1081107</a>
  <script type="text/javascript">
    SimpleTest.waitForExplicitFinish();
    var counter = 0;
    var fcounter = 0;
    var createdCallbackCalled = false;
    function loaded() {
      counter++;
    }
    function failed() {
      fcounter++;
    }
  </script>
  <link rel="import" href="file_element_upgrade.html" onload="loaded()" onerror="failed()"></link>
  <script type="text/javascript">
    is(counter, 1, "Imports are loaded");
    is(fcounter, 0, "No error in imports");
    ok(createdCallbackCalled, "custom element in import has been upgraded");
    SimpleTest.finish();
  </script>
</body>
</html>
